+++ game/chess.py
@@ -22,34 +22,27 @@ class Ajedres:
             raise PiezaInexistente(f"No existe ninguna pieza en la posicion({x}), ({y}).")
         return pieza
     
-    def validaciones(self, movimiento_exitoso):
-        if not movimiento_exitoso:
-            return False  
-
-        estado = self.estado_del_juego()
-
-        if estado == "Empate":
-            return "Empate"
-        elif estado == "Victoria Blanca":
-            return "Victoria Blanca"
-        elif estado == "Victoria Negra":
-            return "Victoria Negra"
-
-      
-        self.cambio_de_turno()
-        return True 
-
-
     def estado_del_juego(self):
-        lista = self.__board__.contar_piezas()
+        piezas = self.__board__.contar_piezas()
+        
+        if piezas == [1, 1]:
+            return "Empate"
+        elif piezas == [1, 0]:
+            return "Victoria Blanca"
+        elif piezas == [0, 1]:
+            return "Victoria Negra"
+    
+        return "En curso"
 
-        if lista == (1,1) :
-            return "Empate" 
 
-        elif lista == (1,0) :
-            return "Victoria Blanca"  
+    def validaciones(self):
+        piezas_alive = self.__board__.contar_piezas()
 
-        elif lista== (0,1):
-            return "Victoria Negra" 
+        if piezas_alive[0] == 1 and piezas_alive[1] == 0:
+            return "Victoria Blanca"
+        elif piezas_alive[0] == 0 and piezas_alive[1] == 1:
+            return "Victoria Negra"
+        elif piezas_alive[0] == 1 and piezas_alive[1] == 1:
+            return "Empate"
 
-        return 
+        return "En curso"